\name{analysedDataset}
\alias{analysedDataset}
\title{Method "analysedDataset"}
\docType{methods}
\description{
    Method \code{analysedDataset} returns subset of the original dataset 
    that was analysed
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns analysed dataset
        }
    }}
\keyword{methods}