\name{analysedDatasetPhenList}
\alias{analysedDatasetPhenList}
\title{Method "analysedDatasetPhenList"}
\docType{methods}
\description{
    Method \code{analysedDatasetPhenList} returns subset of the original dataset 
    that will be analysed
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns analysed dataset
        }
    }}
\keyword{methods}