\name{analysedSubset}
\alias{analysedSubset}
\title{Method "analysedSubset"}
\docType{methods}
\description{
    Method \code{analysedSubset} returns name of the subset that was analysed by RR or FE frameworks
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "htestPhenStat")}}{
            %%  Returns name of the subset
        }
    }}
\keyword{methods}