\name{analysisResults}
\alias{analysisResults}
\title{Method "analysisResults"}
\docType{methods}
\description{
    Method \code{analysisResults} returns analysis results
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns results of the analysis
        }
    }}
\keyword{methods}