\name{batchIn}
\alias{batchIn}
\title{Method "batchIn"}
\docType{methods}
\alias{batchIn-methods}
\alias{batchIn,PhenList-method}
\alias{batchIn,PhenTestResult-method}
\description{
 Method \code{batchIn} returns TRUE/FALSE values depending on the batch column presence/absence in the dataset
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "PhenList")}}{
%%  Returns  TRUE/FALSE values depending on batch column presence/absence
}

\item{\code{signature(obj = "PhenTestResult")}}{
%%  Returns  TRUE/FALSE values depending on batch column presence/absence
}
}}
\keyword{methods}