\name{comparison}
\alias{comparison}
\title{Method "comparison"}
\docType{methods}
\description{
    Method \code{comparison} returns name of the comparison that was performed by RR framework: 
    "High vs Normal/Low" or "Low vs Normal/High"
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "htestPhenStat")}}{
            %%  Returns name of the comparison
        }
    }}
\keyword{methods}