\name{dataset}
\alias{dataset}
\title{Method "dataset"}
\docType{methods}
\description{
    Method \code{dataset} returns data frame stored in the object
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns data frame
        }
    }}
\keyword{methods}