\name{getColumn}
\alias{getColumn}
\title{Method "getColumn"}
\docType{methods}
\description{
    Method \code{getColumn} returns column of interest
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns column of interest
        }
    }}
\keyword{methods}