\name{getColumnBatchAdjusted}
\alias{getColumnBatchAdjusted}
\title{Method "getColumnBatchAdjusted"}
\docType{methods}
\description{
    Method \code{getColumnBatchAdjusted} returns column of interest adjested for batch
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns column of interest adjusted for batch
        }
    }}
\keyword{methods}