\name{getColumnView}
\alias{getColumnView}
\title{Method "getColumnView"}
\docType{methods}
\description{
    Method \code{getColumnView} returns RR or FE frameworks results in a column format
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "htestPhenStat")}}{
            %%  Returns result in a column format
        }
    }}
\keyword{methods}