\name{getColumnWeightBatchAdjusted}
\alias{getColumnWeightBatchAdjusted}
\title{Method "getColumnWeightBatchAdjusted"}
\docType{methods}
\description{
    Method \code{getColumnWeightBatchAdjusted} returns column of interest 
    adjusted for weight and batch
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns column of interest adjusted for batch
        }
    }}
\keyword{methods}