\name{getCountMatrices}
\alias{getCountMatrices}
\title{Method "getCountMatrices"}
\docType{methods}
\description{
    Method \code{getCountMatrices} returns count matrices for FE and RR frameworks
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns count matrices 
        }
    }}
\keyword{methods}