\name{getDataset}
\alias{getDataset}
\title{Method "getDataset"}
\docType{methods}
\description{
    Method \code{getDataset} returns data frame stored in the object
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns data frame
        }
    }}
\keyword{methods}