\name{getGenotypeEffect}
\alias{getGenotypeEffect}
\title{Method "getGenotypeEffect"}
\docType{methods}
\description{
    Method \code{getGenotypeEffect} returns genotype effect and standard error for the linear regression frameworks
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns genotype effect and standard error
        }
    }}
\keyword{methods}