\name{getMatrix}
\alias{getMatrix}
\title{Method "getMatrix"}
\docType{methods}
\description{
    Method \code{getMatrix} returns RR or FE frameworks results in a matrix format
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "htestPhenStat")}}{
            %%  Returns result in a matrix format
        }
    }}
\keyword{methods}