\name{getPercentageMatrix}
\alias{getPercentageMatrix}
\title{Method "getPercentageMatrix"}
\docType{methods}
\description{
    Method \code{getPercentageMatrix} returns percentage matrix calculated out of counts
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "htestPhenStat")}}{
            %%  Returns percentage matrix
        }
    }}
\keyword{methods}