\name{getStat}
\alias{getStat}
\title{Method "getStat"}
\docType{methods}
\description{
    Method \code{getStat} returns simple statistics about variables in the dataset
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns simple statistics about variables in the dataset
        }
    }}
\keyword{methods}