\name{getVariable}
\alias{getVariable}
\title{Method "getVariable"}
\docType{methods}
\description{
    Method \code{getVariable} returns dependent variable name
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns dependent variable name 
        }
    }}
\keyword{methods}