\name{getVariables}
\alias{getVariables}
\title{Method "getVariables"}
\docType{methods}
\description{
    Method \code{getVariables} returns names of variables in the dataset
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns names of the variables in the dataset
        }
    }}
\keyword{methods}