\name{hemiGenotype}
\alias{hemiGenotype}
\title{Method "hemiGenotype"}
\docType{methods}
\description{
    Method \code{hemiGenotype} returns hemi genotype if defined
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns hemi genotype
        }
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns hemi genotype
        }
    }}
\keyword{methods}