\name{matrixCount}
\alias{matrixCount}
\title{Method "matrixCount"}
\docType{methods}
\description{
    Method \code{matrixCount} returns matrix of counts that was created by RR or FE framework.
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "htestPhenStat")}}{
            %%  Returns matrix of counts
        }
    }}
\keyword{methods}