\name{method}
\alias{method}
\title{Method "method"}
\docType{methods}
\description{
    Method \code{method} returns method abbriviation, for example: "RR", "MM".
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns method abbriviation
        }
    }}
\keyword{methods}