\name{methodText}
\alias{methodText}
\title{Method "methodText"}
\docType{methods}
\description{
    Method \code{methodText} returns method's name, for example: "Reference Range Plus", "Linear Mixed Model".
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns method's name
        }
    }}
\keyword{methods}