\name{multipleBatches}
\alias{multipleBatches}
\title{Method "multipleBatches"}
\docType{methods}
\description{
    Method \code{multipleBatches} returns TRUE if there are multiple batches in the dataset, FALSE otherwise
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns TRUE if there are multiple batches
        }
    }}
\keyword{methods}