\name{noSexes}
\alias{noSexes}
\title{Method "noSexes"}
\docType{methods}
\description{
    Method \code{noSexes} returns number of sex levels: 1/2
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns number of sex levels
        }
    }}
\keyword{methods}