\name{parameters}
\alias{parameters}
\title{Method "parameters"}
\docType{methods}
\description{
    Method \code{parameters} returns parameters specific for apllied method/dataset combination
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns parameters specific for apllied method/dataset combination
        }
    }}
\keyword{methods}