\name{pvalue}
\alias{pvalue}
\title{Method "pvalue"}
\docType{methods}
\description{
    Method \code{pvalue} returns p-value that was calculate by RR or FE framework.
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "htestPhenStat")}}{
            %%  Returns p-value of the comparison
        }
    }}
\keyword{methods}