\name{refGenotype}
\alias{refGenotype}
\title{Method "refGenotype"}
\docType{methods}
\description{
    Method \code{refGenotype} returns reference genotype
}
\section{Methods}{
    \describe{      
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns reference genotype
        }

        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns reference genotype
        }
    }}
\keyword{methods}