\name{setBatch}
\alias{setBatch}
\title{Method "setBatch"}
\docType{methods}
\description{
    Method \code{setBatch} sets batch column in the dataset
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Sets batch column
        }
    }}
\keyword{methods}