\name{setGenotype}
\alias{setGenotype}
\title{Method "setGenotype"}
\docType{methods}
\description{
    Method \code{setGenotype} sets genotype column in the dataset
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Sets genotype column
        }
    }}
\keyword{methods}