\name{setMissingValue}
\alias{setMissingValue}
\title{Method "setMissingValue"}
\docType{methods}
\description{
    Method \code{setMissingValue} sets missing value in the dataset
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Sets missing value
        }
    }}
\keyword{methods}