\name{setSex}
\alias{setSex}
\title{Method "setSex"}
\docType{methods}
\description{
    Method \code{setSex} sets sex column in the dataset
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Sets sex column
        }
    }}
\keyword{methods}