\name{setWeight}
\alias{setWeight}
\title{Method "setWeight"}
\docType{methods}
\description{
    Method \code{setWeight} sets weight column in the dataset
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Sets weight column
        }
    }}
\keyword{methods}