\name{subsetText}
\alias{subsetText}
\title{Method "subsetText"}
\docType{methods}
\description{
    Method \code{subsetText} returns full name of subset that was analysed by RR or FE framework:
    Males only, Females only, All.
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "htestPhenStat")}}{
            %%  Returns full name of subset
        }
    }}
\keyword{methods}