\name{testGenotype}
\alias{testGenotype}
\title{Method "testGenotype"}
\docType{methods}
\description{
    Method \code{testGenotype} returns test genotype
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns test genotype
        }
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns test genotype
        }
    }}
\keyword{methods}