\name{transformation}
\alias{transformation}
\title{Method "transformation"}
\docType{methods}
\description{
    Method \code{transformation} returns transformation values: lamda=value, scaleShift=value 
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns transformation values
        }
    }}
\keyword{methods}