\name{transformationJSON}
\alias{transformationJSON}
\title{Method "transformationJSON"}
\docType{methods}
\description{
    Method \code{transformationJSON} returns transformation values as JSON string
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns transformation values as one string
        }
    }}
\keyword{methods}