\name{transformationText}
\alias{transformationText}
\title{Method "transformationText"}
\docType{methods}
\description{
    Method \code{transformationText} returns transformation values as one string
}
\section{Methods}{
    \describe{       
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns transformation values as one string
        }
    }}
\keyword{methods}