\name{weightIn}
\alias{weightIn}
\docType{methods}
\alias{weightIn-methods}
\alias{weightIn,PhenList-method}
\alias{weightIn,PhenTestResult-method}
\title{Method "weightIn"}
\description{
    Returns TRUE if the weight is in the dataset, FALSE otherwise
}
\section{Methods}{
    \describe{
        
        \item{\code{signature(obj = "PhenList")}}{
            %%  Returns  TRUE/FALSE values depending on weight column presence/absence
        }
        
        \item{\code{signature(obj = "PhenTestResult")}}{
            %%  Returns  TRUE/FALSE values depending on weight column presence/absence
        }
    }}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
